fix(objectToArraylist) : corrected build issues
